\name{rinbotools-package}
\alias{rinbotools-package}
\alias{rinbotools}
\docType{package}
\title{
\packageTitle{rinbotools}
}
\description{
\packageDescription{rinbotools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rinbotools}
\packageIndices{rinbotools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rinbotools}

Maintainer: \packageMaintainer{rinbotools}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
